requirements.txt
grpcio==1.51.3
grpcio-tools==1.51.3
jsonpickle==3.0.1
numpy==1.24.2
pika==1.3.1
redis==4.5.1
scipy==1.10.1
matplotlib==3.4.3
